exeptions and set core started

+++ signals.cpp
@@ -13,20 +13,24 @@ void ctrlCHandler(int sig_num)
   Command *current_command = smash.getCurrentCommand();
   if (current_command != nullptr)
   {
-    int pid = current_command->getProcessId();
-    if (kill(pid, SIGKILL) == -1)
+    std::string cmd_line(current_command->getCmdL());
+    if (!(isPipe(cmd_line) || isRedirect(cmd_line) || isSterrPipe(cmd_line) || isAppendRedirect(cmd_line)))
     {
-      // לזרוק שגיאה שkill לא עבדה
-      // perror("...kill...");
-    }
-    else
-    {
-      if (current_command->getJobId() != -1)
+      int pid = current_command->getProcessId();
+      if (kill(pid, SIGKILL) == -1)
       {
-        smash.removeJob(current_command->getJobId());
+        // לזרוק שגיאה שkill לא עבדה
+        // perror("...kill...");
       }
-      smash.setCurrentCommand(nullptr);
-      std::cout << "smash: process " << pid << " was killed" << std::endl;
+      else
+      {
+        if (current_command->getJobId() != -1)
+        {
+          smash.removeJob(current_command->getJobId());
+        }
+        smash.setCurrentCommand(nullptr);
+        std::cout << "smash: process " << pid << " was killed" << std::endl;
+      }
     }
   }
 }
@@ -38,17 +42,21 @@ void ctrlZHandler(int sig_num)
   Command *current_command = smash.getCurrentCommand();
   if (current_command != nullptr)
   {
-    int pid = current_command->getProcessId();
-    if (kill(pid, SIGSTOP) == -1)
+    std::string cmd_line(current_command->getCmdL());
+    if (!(isPipe(cmd_line) || isRedirect(cmd_line) || isSterrPipe(cmd_line) || isAppendRedirect(cmd_line)))
     {
-      // לזרוק שגיאה שkill לא עבדה
-      // perror("...kill...");
-    }
-    else
-    {
-      smash.addJob(current_command, true);
-      smash.setCurrentCommand(nullptr);
-      std::cout << "smash: process " << pid << " was stopped" << std::endl;
+      int pid = current_command->getProcessId();
+      if (kill(pid, SIGSTOP) == -1)
+      {
+        // לזרוק שגיאה שkill לא עבדה
+        // perror("...kill...");
+      }
+      else
+      {
+        smash.addJob(current_command, true);
+        smash.setCurrentCommand(nullptr);
+        std::cout << "smash: process " << pid << " was stopped" << std::endl;
+      }
     }
   }
 }
